# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: List all active subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject by ID
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Chapters ──────────────────────────────────────────────────────────
...
  /gate/check:
    post:
      operationId: checkGate
      tags: [progress]
      summary: Check if a gate/step is accessible for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateCheckInput"
      responses:
        "200":
          description: Gate status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateCheckResult"
        "403":
          description: Prerequisites not met
  /progress/topic/{topicId}:
    get:
      operationId: getTopicProgress
      tags: [progress]
      summary: Get user progress for a specific topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Topic progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicProgress"
  /progress/lecture-click:
    post:
      operationId: recordLectureClick
      tags: [progress]
      summary: Record that user clicked a lecture link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureClickInput"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Exams ─────────────────────────────────────────────────────────────
...
  /notes/{noteId}:
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── B2 Storage ────────────────────────────────────────────────────────
  /b2/upload-url:
    post:
      operationId: getUploadUrl
      tags: [storage]
      summary: Get presigned B2 URL for PDF upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResult"
        "403":
          description: Gate not unlocked or quota exceeded
  /b2/download-url:
    post:
      operationId: getDownloadUrl
      tags: [storage]
      summary: Get presigned B2 URL for PDF download
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadUrlInput"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResult"
  /b2/profile-upload-url:
    post:
      operationId: getProfileUploadUrl
      tags: [storage]
      summary: Get presigned B2 URL for profile photo upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUploadInput"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResult"
  /b2/profile-download-url:
    post:
      operationId: getProfileDownloadUrl
      tags: [storage]
      summary: Get presigned B2 URL for profile photo download
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileDownloadInput"
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResult"
...
  /pomodoro/sessions:
    get:
      operationId: getPomodoroSessions
      tags: [pomodoro]
      summary: Get user's pomodoro session history
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PomodoroSession"
    post:
      operationId: createPomodoroSession
      tags: [pomodoro]
      summary: Log a completed pomodoro session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroSessionInput"
      responses:
        "201":
          description: Session logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSession"
...
  /external-tests:
    get:
      operationId: getExternalTests
      tags: [tests]
      summary: Get user's external test log
      responses:
        "200":
          description: External tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExternalTest"
    post:
      operationId: createExternalTest
      tags: [tests]
      summary: Log an external test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalTestInput"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalTest"
  /external-tests/{testId}:
    delete:
      operationId: deleteExternalTest
      tags: [tests]
      summary: Delete an external test entry
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Admin ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Subject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        icon_url:
          type: ["string", "null"]
        is_active:
          type: boolean
      required: [id, title, order_index, is_active]
    SubjectInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        icon_url:
          type: ["string", "null"]
        is_active:
          type: boolean
      required: [title, order_index]
    SubjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order_index:
          type: integer
        is_active:
          type: boolean
...
    GateCheckInput:
      type: object
      properties:
        topic_id:
          type: string
        step:
          type: string
          enum: [lecture, lecture_quiz, dpp, pyqs, topic_test]
      required: [topic_id, step]
    GateCheckResult:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: ["string", "null"]
      required: [allowed]
    TopicProgress:
      type: object
      properties:
        topic_id:
          type: string
        lecture_clicked:
          type: boolean
        lecture_quiz_passed:
          type: boolean
        lecture_quiz_score:
          type: ["number", "null"]
        dpp_completed:
          type: boolean
        dpp_score:
          type: ["number", "null"]
        pyqs_completed:
          type: boolean
        pyqs_score:
          type: ["number", "null"]
        topic_test_passed:
          type: boolean
        topic_test_score:
          type: ["number", "null"]
        topic_complete:
          type: boolean
      required: [topic_id, lecture_clicked, lecture_quiz_passed, dpp_completed, pyqs_completed, topic_test_passed, topic_complete]
    LectureClickInput:
      type: object
      properties:
        lecture_id:
          type: string
        topic_id:
          type: string
      required: [lecture_id, topic_id]

    # ── Exam ─────────────────────────────────────────────────────────────
...
    UploadUrlInput:
      type: object
      properties:
        chapter_id:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
      required: [chapter_id, filename, content_type, size_bytes]
    UploadUrlResult:
      type: object
      properties:
        upload_url:
          type: string
        storage_path:
          type: string
        expires_at:
          type: string
      required: [upload_url, storage_path, expires_at]
    DownloadUrlInput:
      type: object
      properties:
        storage_path:
          type: string
      required: [storage_path]
    DownloadUrlResult:
      type: object
      properties:
        download_url:
          type: string
        expires_at:
          type: string
      required: [download_url, expires_at]
    ProfileUploadInput:
      type: object
      properties:
        content_type:
          type: string
        size_bytes:
          type: integer
      required: [content_type, size_bytes]
    ProfileDownloadInput:
      type: object
      properties:
        user_id:
          type: string
      required: [user_id]
...
    PomodoroSession:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        duration_seconds:
          type: integer
        topic_context:
          type: ["string", "null"]
        start_time:
          type: string
        end_time:
          type: string
      required: [id, user_id, duration_seconds, start_time, end_time]
    PomodoroSessionInput:
      type: object
      properties:
        duration_seconds:
          type: integer
        topic_context:
          type: ["string", "null"]
        start_time:
          type: string
        end_time:
          type: string
      required: [duration_seconds, start_time, end_time]
...
    ExternalTest:
      type: object
      properties:
        id:
          type: string
        exam_name:
          type: string
        exam_date:
          type: string
        score_obtained:
          type: number
        total_marks:
          type: number
        percentile:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, exam_name, exam_date, score_obtained, total_marks, created_at]
    ExternalTestInput:
      type: object
      properties:
        exam_name:
          type: string
        exam_date:
          type: string
        score_obtained:
          type: number
        total_marks:
          type: number
        percentile:
          type: ["number", "null"]
        rank:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [exam_name, exam_date, score_obtained, total_marks]

    # ── Admin ─────────────────────────────────────────────────────────────
...